pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "twitter_data_pipeline"
authors = ["Timothy Cameron <timiancam@gmail.com>"]
version = "1.0"
description = "N/A"
readme = "README.md"
keywords = ["Spotify", "Airflow", "Apache", "AWS", "EC2", "S3", "Spotipy", "Python", "Pandas"]
license = "MIT"
repository = "https://github.com/timiancam/Spotify_pipeline_EC2_S3_Apache"
packages = [
    { include = "Spotify_pipeline_EC2_S3_Apache" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pandas = "^2.1.3"
spotipy = "^2.23.0"
python-decouple = "3.6"
s3fs = "^2024.2.0"
